+++ htmldoc/ps-pdf.cxx
@@ -4391,12 +4391,12 @@ parse_doc(tree_t *t,		/* I - Tree to parse */
 	    *needspace = 1;
           }
 
-          *left += 36.0f;
+          *left += (float)PreIndent;
           *x    = *left;
 
           parse_pre(t, *left, *right, *bottom, *top, x, y, page, *needspace);
 
-          *left      -= 36.0f;
+          *left      -= (float)PreIndent;
           *x         = *left;
           *needspace = 1;
           break;